+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/actions/counters/ResumeLiveCounterSessionAction.kt
@@ -5,12 +5,12 @@ import com.github.rafaelldi.diagnosticsclientplugin.dialogs.MonitoringTimerDialo
 import com.github.rafaelldi.diagnosticsclientplugin.generated.LiveCounterSession
 import com.github.rafaelldi.diagnosticsclientplugin.generated.diagnosticsHostModel
 import com.github.rafaelldi.diagnosticsclientplugin.services.counters.LiveCounterSessionController
-import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.tabs.CounterMonitoringTab
-import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.tabs.CounterMonitoringTab.Companion.COUNTER_MONITORING_TAB
+import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.tabs.LiveCounterSessionTab
+import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.tabs.LiveCounterSessionTab.Companion.COUNTER_MONITORING_TAB
 import com.intellij.openapi.project.Project
 import com.jetbrains.rider.projectView.solution
 
-class ResumeLiveCounterSessionAction : ResumeLiveSessionAction<LiveCounterSession, CounterMonitoringTab>() {
+class ResumeLiveCounterSessionAction : ResumeLiveSessionAction<LiveCounterSession, LiveCounterSessionTab>() {
     override val tabDatKey = COUNTER_MONITORING_TAB
 
     override fun resumeSession(pid: Int, project: Project) {