+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/actions/traces/PauseLiveTraceSessionAction.kt
@@ -4,12 +4,12 @@ import com.github.rafaelldi.diagnosticsclientplugin.actions.common.PauseLiveSess
 import com.github.rafaelldi.diagnosticsclientplugin.generated.LiveTraceSession
 import com.github.rafaelldi.diagnosticsclientplugin.generated.diagnosticsHostModel
 import com.github.rafaelldi.diagnosticsclientplugin.services.traces.LiveTraceSessionController
-import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.tabs.TraceMonitoringTab
-import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.tabs.TraceMonitoringTab.Companion.TRACE_MONITORING_TAB
+import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.tabs.LiveTraceSessionTab
+import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.tabs.LiveTraceSessionTab.Companion.TRACE_MONITORING_TAB
 import com.intellij.openapi.project.Project
 import com.jetbrains.rider.projectView.solution
 
-class PauseLiveTraceSessionAction : PauseLiveSessionAction<LiveTraceSession, TraceMonitoringTab>() {
+class PauseLiveTraceSessionAction : PauseLiveSessionAction<LiveTraceSession, LiveTraceSessionTab>() {
     override val tabDatKey = TRACE_MONITORING_TAB
 
     override fun pauseSession(pid: Int, project: Project) {