+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/services/gc/PersistentGcEventSessionController.kt
@@ -1,8 +1,8 @@
 package com.github.rafaelldi.diagnosticsclientplugin.services.gc
 
-import com.github.rafaelldi.diagnosticsclientplugin.common.collectionSessionAlreadyExists
-import com.github.rafaelldi.diagnosticsclientplugin.common.collectionSessionFinished
-import com.github.rafaelldi.diagnosticsclientplugin.common.collectionSessionStarted
+import com.github.rafaelldi.diagnosticsclientplugin.common.persistentSessionAlreadyExists
+import com.github.rafaelldi.diagnosticsclientplugin.common.persistentSessionFinished
+import com.github.rafaelldi.diagnosticsclientplugin.common.persistentSessionStarted
 import com.github.rafaelldi.diagnosticsclientplugin.dialogs.GcEventModel
 import com.github.rafaelldi.diagnosticsclientplugin.dialogs.StoppingType
 import com.github.rafaelldi.diagnosticsclientplugin.generated.PersistentGcEventSession
@@ -40,8 +40,8 @@ class PersistentGcEventSessionController(project: Project) :
         return PersistentGcEventSession(duration, filePath)
     }
 
-    override fun sessionAlreadyExists(pid: Int) = collectionSessionAlreadyExists(GC_EVENTS, pid, project)
-    override fun sessionStarted(pid: Int) = collectionSessionStarted(GC_EVENTS, pid, project)
+    override fun sessionAlreadyExists(pid: Int) = persistentSessionAlreadyExists(GC_EVENTS, pid, project)
+    override fun sessionStarted(pid: Int) = persistentSessionStarted(GC_EVENTS, pid, project)
     override fun sessionFinished(pid: Int, session: PersistentGcEventSession) =
-        collectionSessionFinished(GC_EVENTS, pid, session.filePath, true, project)
+        persistentSessionFinished(GC_EVENTS, pid, session.filePath, true, project)
 }