+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/toolWindow/tabs/LiveTraceSessionTab.kt
@@ -20,7 +20,7 @@ import java.awt.BorderLayout
 import java.text.SimpleDateFormat
 import javax.swing.JPanel
 
-class TraceMonitoringTab(
+class LiveTraceSessionTab(
     override val pid: Int,
     session: LiveTraceSession,
     private val manager: TraceTabManager,
@@ -28,7 +28,7 @@ class TraceMonitoringTab(
     lt: Lifetime
 ) : SimpleToolWindowPanel(false), MonitoringTab, Disposable {
     companion object {
-        val TRACE_MONITORING_TAB: DataKey<TraceMonitoringTab> =
+        val TRACE_MONITORING_TAB: DataKey<LiveTraceSessionTab> =
             DataKey.create("DiagnosticsClient.ToolWindow.TraceMonitoringTab")
 
         private val HTTP_OUTPUT = ConsoleViewContentType("HTTP_OUTPUT", ConsoleHighlighter.BLUE)