+++ src/rider/main/kotlin/icons/DiagnosticsClientIcons.kt
@@ -8,6 +8,4 @@ object DiagnosticsClientIcons {
     val ToolWindow = IconLoader.getIcon("/icons/diagnosticsClientTool.svg", javaClass)
     @JvmField
     val Counters = IconLoader.getIcon("/icons/counters.svg", javaClass)
-    @JvmField
-    val Traces = IconLoader.getIcon("/icons/traces.svg", javaClass)
 }